refactor: Add index.ts in icons
