+++ src/shared/icons/index.ts
@@ -1,0 +1,12 @@
+export { default as BusIcon } from './bus';
+export { default as CameraIcon } from './camera';
+export { default as CheckIcon } from './check';
+export { default as ChevronLeftIcon } from './chevron-left';
+export { default as CloseIcon } from './close';
+export { default as CloudIcon } from './cloud';
+export { default as LiftIcon } from './lift';
+export { default as RainIcon } from './rain';
+export { default as ShareIcon } from './share';
+export { default as SnowIcon } from './snow';
+export { default as SunIcon } from './sun';
+export { default as VoteIcon } from './vote';
