+++ src/widgets/discovery-detail/ui/vote-dialog.tsx
@@ -1,5 +1,4 @@
 import { useCallback, useState } from 'react';
-import CheckIcon from '@/shared/icons/check';
 import { cn } from '@/shared/lib';
 import {
   Dialog,
@@ -11,6 +10,7 @@ import {
   DialogTitle,
   DialogTrigger,
 } from '@/shared/ui/dialog';
+import { CheckIcon } from '@/shared/icons';
 
 interface VoteDialogProps {
   trigger: React.ReactNode;
